+++ src/OpenTelemetry/ProviderExtensions.cs
@@ -57,7 +57,23 @@ namespace OpenTelemetry
         /// <returns><see cref="Resource"/>if found otherwise <see cref="Resource.Empty"/>.</returns>
         public static Resource GetDefaultResource(this BaseProvider baseProvider)
         {
-            return ResourceBuilder.CreateDefault().Build();
+            var builder = ResourceBuilder.CreateDefault();
+            builder.ServiceProvider = GetServiceProvider(baseProvider);
+            return builder.Build();
+        }
+
+        internal static IServiceProvider GetServiceProvider(this BaseProvider baseProvider)
+        {
+            if (baseProvider is TracerProviderSdk tracerProviderSdk)
+            {
+                return tracerProviderSdk.ServiceProvider;
+            }
+            else if (baseProvider is MeterProviderSdk meterProviderSdk)
+            {
+                return meterProviderSdk.ServiceProvider;
+            }
+
+            return null;
         }
 
         internal static Action GetObservableInstrumentCollectCallback(this BaseProvider baseProvider)