+++ src/OpenTelemetry/CHANGELOG.md
@@ -8,6 +8,11 @@
   ([#3760](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3760),
   [#3776](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3776))
 
+* Added dependency injection support in the `ResourceBuilder` class and added
+  support for loading environment variables from `IConfiguration` for the
+  `AddEnvironmentVariableDetector` extension
+  ([#3782](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3782))
+
 ## 1.4.0-beta.2
 
 Released 2022-Oct-17