+++ src/OpenTelemetry/CHANGELOG.md
@@ -11,7 +11,8 @@
 * Added dependency injection support in the `ResourceBuilder` class and added
   support for loading environment variables from `IConfiguration` for the
   `AddEnvironmentVariableDetector` extension
-  ([#3782](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3782))
+  ([#3782](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3782),
+  [#3798](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3798))
 
 ## 1.4.0-beta.2