+++ src/OpenTelemetry/CHANGELOG.md
@@ -8,6 +8,19 @@
   ([#3760](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3760),
   [#3776](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3776))
 
+* Added dependency injection support in the `ResourceBuilder` class and added
+  support for loading environment variables from `IConfiguration` for the
+  `AddEnvironmentVariableDetector` extension
+  ([#3782](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3782),
+  [#3798](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3798))
+
+* Breaking: MetricPoint API to retrieve Histogram Min, Max changed. The existing
+  pattern of checking if Min/Max is available with `HasMinMax()` and then
+  retrieving the same using `GetHistogramMin()`, `GetHistogramMax()` is replaced
+  with a single API `TryGetHistogramMinMaxValues(out double min, out double
+  max)`.
+  ([#3822](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3822))
+
 ## 1.4.0-beta.2
 
 Released 2022-Oct-17