+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fix instrument naming enforcement implementation to match the spec.
+  ([#3821](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3821))
+
 * Added support for loading environment variables from `IConfiguration` when
   using the `MetricReaderOptions` & `BatchExportActivityProcessorOptions`
   classes.
@@ -13,6 +16,13 @@
   `AddEnvironmentVariableDetector` extension
   ([#3782](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3782),
   [#3798](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3798))
+
+* Breaking: MetricPoint API to retrieve Histogram Min, Max changed. The existing
+  pattern of checking if Min/Max is available with `HasMinMax()` and then
+  retrieving the same using `GetHistogramMin()`, `GetHistogramMax()` is replaced
+  with a single API `TryGetHistogramMinMaxValues(out double min, out double
+  max)`.
+  ([#3822](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3822))
 
 ## 1.4.0-beta.2