+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.4.0-beta.3
+
+Released 2022-Nov-07
+
 * Fix instrument naming enforcement implementation to match the spec.
   ([#3821](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3821))